split the app.py into a module init and a routes file
