+++ app/__init__.py
@@ -1,2 +1,10 @@
 # flake8: noqa
-from .app import manager
+from flask import Flask
+from flask.ext.script import Manager
+from app import routes
+
+app = Flask(__name__)
+app.config.from_object('config')
+manager = Manager(app)
+db.init(app, manager)
+routes.routes(app)
